fix(uploads): bind checksum header into presigned URL, fix race + bigint
Addresses CI test failure and review feedback on #497.

CI: `presignUploadUrl` unit test failed because
`x-amz-checksum-sha256` never appeared in the signed URL's
`X-Amz-SignedHeaders`. SDK v3's flexible-checksums middleware
only translates the `ChecksumSHA256` command field into a request
header when it can compute against a real body; presigning has no
body, so the middleware skipped it and the signer never saw the
header. Adding a per-command middleware at the `build` step
injects the header directly so the signer binds it. The
`@smithy/protocol-http` `HttpRequest.isInstance` guard keeps the
narrowing safe; it's added as a direct dep even though it was
already a transitive of the AWS SDK.

bigint for `declared_size` (gemini-code-assist high): `integer`
caps at 2.14 GB. S3 supports up to 5 TB; uploads above 2 GB would
have crashed on insert. Updated migration + drizzle schema.

Race in lazy client init (gemini-code-assist medium): the previous
`_awsClient` cache was set after the awaited build resolved, so
concurrent first-callers each kicked off their own
`buildAwsS3Client()`. Caching the promise instead of the resolved
client lets all callers in the build window share a single
credential-resolution round trip.

Revert `S3_CREDENTIALS_PROVIDER = "env"` in test env (codex
P2): the existing `s3.test.ts` cases verify ECS/IMDS precedence
under the default `auto` mode. Forcing `env` globally made them
exit early on static credentials before ever calling the mocked
runtime fetches, masking auto-mode regressions. The static MinIO
credentials are enough for sign-only unit tests under `auto` mode
on the `http://localhost:9000` endpoint.

+++ apps/api/drizzle/20260529160000_presigned_upload_staging/migration.sql
@@ -12,7 +12,7 @@ CREATE TABLE "pending_uploads" (
   "purpose_data" jsonb NOT NULL,
   "declared_name" varchar(255) NOT NULL,
   "declared_mime" varchar(255) NOT NULL,
-  "declared_size" integer NOT NULL,
+  "declared_size" bigint NOT NULL,
   "declared_sha256" varchar(64) NOT NULL,
   "status" text DEFAULT 'pending' NOT NULL,
   "finalized_result" jsonb,
